Add db_uri explanation to the readme
I went round the houses because I thought that the command line option 'db' would be the right key to use in the config yaml. In fact it is db_uri from reading the code. Otherwise fails with Sequel::Database.connect takes either a Hash or a String, given: nil (Sequel::Error)